Initial stab at returning stuff based on object ID
This almost certainly doesn’t work, but should give you an idea of
where I was thinking of going with making requests. The idea is that
this FB_Media class extends the FacebookRequest class and then maps the
variety of fields that you get back when requesting an object ID to
some standard things.

Usage, getting an image in this example:
$fb_object = new FB_Media;
$image = $fb_object->get_images();

Once you’ve got that object, you can query its other methods, such as

$comments = $fb_object->get_comments():